+++ Changelog.md
@@ -1,4 +1,5 @@
 ## 0.0.67
++ Added `beta.LiteRTInterpreterMetadata` class to compile LiteRT `Interpreter` instances for inference.
 + Added support for specifying JSON list and dictionary arguments in `muna predict` CLI command.
 
 ## 0.0.66